Impl pull action.

+++ src/container_config.rs
@@ -1,6 +1,6 @@
 use crate::constants::{NEOVIM_IMAGE_PREFIX, NEOVIM_IMAGE_TAG};
 pub struct ContainerImageConfig {
-    pub uid: u32,
+    pub uid: u16,
 }
 
 pub fn image_name(config: ContainerImageConfig) -> String {
